fixed login signup;added dropdown menu when logged
