+++ package.json
@@ -74,8 +74,10 @@
 	},
 	"dependencies": {
 		"axios": "^1.4.0",
+		"bootstrap": "^5.3.1",
 		"prop-types": "^15.6.1",
 		"react": "^16.8.4",
+		"react-bootstrap": "^2.8.0",
 		"react-dom": "^16.8.4",
 		"react-polyfills": "0.0.1",
 		"react-router-dom": "^6.3.0"
